tests/wpt/web-platform-tests/dom/nodes/NodeList-Iterable.html

<!doctype html>
<meta charset="utf-8">
<title></title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
	<p>p1</p>
	<p>p2</p>
	<p>p3</p>
	<p>p4</p>
	<p>p5</p>
<script>
	function testIterable(nodes) {
		test(function() {
			for(var i=0; i<nodes.length; ++i)
				assert_not_equals(nodes[i], undefined);
		}, "Iterating nodeList via indexing.");

		test(function() {
			assert_equals(nodes[nodes.length], undefined);
		}, "Expect nodeList[length] to be undefined.");

		test(function() {
			for(var i=0; i<nodes.length; ++i){
				assert_equals(nodes[i], nodes.item(i));
			}
		}, "NodeList indexing & item method equality.");
	}

	var paragraphs, nothing;
	setup(function() {
		paragraphs = document.querySelectorAll('p');
		nothing = document.querySelectorAll('span');
	})
	testIterable(paragraphs);
	testIterable(nothing);
</script>
